Added examples, corrected tests.
Added requirement of GeneNet, WGCNA, parmigene, and space. This can be reevaluated in the future, but it was nice to use those packages during the examples knowing that they'd be available. Also, it would likely be more convenient for the user to have them all installed automatically. The only downside would be the increase in wait-time to run the package after opening all of the other dependencies.